renamed mbuild lib dir for CLI usability
